Convert the result of the sqlite plugin to a javascript array
The original assumption was that the sqlite plugin would return an array, which
we would then use to invoke the callback.

However, it turns out that sqlite actually returns an iterator. So we convert
the interator to a list, simultaneously formatting the timestamp to a user
readable format, and then return it.

It is also possible to return the iterator directly, which might actually be
more performant since we have to step through the list anyway to append it to
the entries, and it is wasteful to step through the list twice. But let's do
that as a performance optimization later.