Fixed the code to return entries by index
There is now a separate function to return the max index, and the main code
returns all entries that are less than the index since we display values in
reverse chron order.

Also fixed the error handlers to be as per sqlite plugin expectations - namely
that they take two arguments, the second of which is the error, and they return
false.