Re-open the database if needed on iOS
For some reason, although we call the same logger init code on both iOS and
android, the iOS logger sometimes ends up unopened. We found this in testing -
it does not appear to happen on android and it did not happen in the previous
standalone app.

Fixing this for now to re-open if it is closed. Need to figure out why it is
closed in the first place, but running out of time for that now. Maybe a
service would be appropriate? It can also promisify everything...