Move the init code to the initialize methods
And define the plugin to be loaded onload.
It is now started as part of the initial set of plugins, and doesn't take that long either.

2016-02-11 23:39:02.430 emission[49903:17135447] [CDVTimer][handleopenurl] 0.059009ms
2016-02-11 23:39:02.431 emission[49903:17135447] [CDVTimer][intentandnavigationfilter] 0.930965ms
2016-02-11 23:39:02.432 emission[49903:17135447] [CDVTimer][gesturehandler] 0.046968ms
2016-02-11 23:39:02.439 emission[49903:17135447] [CDVTimer][file] 7.634044ms
2016-02-11 23:39:02.462 emission[49903:17135447] [CDVTimer][splashscreen] 22.060990ms
2016-02-11 23:39:02.469 emission[49903:17135447] [CDVTimer][statusbar] 6.465018ms
2016-02-11 23:39:02.469 emission[49903:17135447] [CDVTimer][keyboard] 0.137031ms
2016-02-11 23:39:02.469 emission[49903:17135447] UnifiedLogger:pluginInitialize singleton -> initialize native DB
2016-02-11 23:39:02.471 emission[49903:17135447] [CDVTimer][unifiedlogger] 2.453983ms
2016-02-11 23:39:02.472 emission[49903:17135447] [CDVTimer][TotalPluginStartup] 41.270971ms

We could conceivably remove the db variable init from there and put it into the
actual functions and close them later. But it seems wasteful to open it for every transaction.
Let's wait and see how best to optimize that.