Add in support for truncating the log
Right now, the truncation amount is a hardcoded constant that is the same on
both android and iOS (30 days). We can move this to a config option if we need
to.

As part of the change, did a bit of restructuring - creating a new method for
the current time secs in android and for executing an arbitrary delete command
in iOS.

Also removed a println statement that said "Returning logger" for every log
statement and cluttered up the logs.