Add support for displaying notifications
Since we repurposed notification display for logging, and then turned it off
because it was too distracting, we didn't have a way to just display
notifications. This change adds a new method to display notifications, and
refactors existing code to invoke it when necessary. This seems to be a fix for
https://github.com/e-mission/e-mission-data-collection/issues/113