Add a new notification helper that takes the pending intent as an argument
Instead of creating one. Since we do our checks primarily in the background,
the goal is to create a notification that opens the resolution pending
intents when the notification is clicked.

Unfortunately, this doesn't seem to work right now, but it is good to add
support for it in case it does...