Unify the logger database as well
In https://github.com/e-mission/cordova-usercache/pull/20, and follow on commits
https://github.com/e-mission/cordova-usercache/pull/21 and
https://github.com/e-mission/cordova-usercache/pull/22, we ensured that all
accesses to the usercache went through the same handle.

It looks like sqlite on android (and maybe iOS) doesn't really support
multi-threaded access to the same database because it throws a database locked
error instead.
https://github.com/e-mission/e-mission-data-collection/issues/124

We haven't run into this as much with the logs because accessing the logs is
hidden in the developer zone, but it could happen! So let's unify accesses to
the logs as well. We should be able to remove the sqlite plugin once we remove
the stats code.