Add a small icon for the notifications from the native code
Without the smallicon the notifications are not displayed.
The previous smallicon was just the icon, which is not displayed on the
notifications and is replaced with a gray/white circle instead.

This adds an icon at various densities, copies them in the plugin.xml and
changes to code to refer to them.